backend/requirements.txt
flask
flask-cors
google-auth-oauthlib
google-api-python-client
python-dotenv
gunicorn